safari-600.7-branch/LayoutTests/http/tests/notifications/legacy/show-expected.txt

This test looks for the dispatch of the onshow event handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.webkitNotifications.checkPermission() is 1
PASS window.webkitNotifications.createNotification('', 'title 1', 'body'); threw exception Error: SecurityError: DOM Exception 18.
Granting permission to http://127.0.0.1:8000
PASS window.webkitNotifications.checkPermission() is 0
PASS ondisplay event handler was called
PASS this.ondisplay is this.onshow
Denying permission to http://127.0.0.1:8000
PASS window.webkitNotifications.checkPermission() is 2
PASS window.webkitNotifications.createNotification('', 'title 3', 'body'); threw exception Error: SecurityError: DOM Exception 18.
Resetting all permissions
PASS window.webkitNotifications.checkPermission() is 1
PASS successfullyParsed is true

TEST COMPLETE
